backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
youtube-transcript-api>=1.2.0
google-cloud-storage>=2.14.0
google-cloud-firestore>=2.14.0
google-cloud-discoveryengine>=0.13.0
python-dotenv>=1.0.0
scrapetube>=2.5.0
pydantic-settings>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-multipart>=0.0.6
